Move hooks folder
