+++ src/ui/components/seekbar/SeekBar.tsx
@@ -2,9 +2,9 @@ import React, { useContext, useState } from 'react';
 import { type LayoutChangeEvent, StyleProp, View, ViewStyle } from 'react-native';
 import { PlayerContext, UiContext } from '../util/PlayerContext';
 import { Slider } from '@miblanchard/react-native-slider';
-import { useDuration } from '../hooks/useDuration';
-import { useSeekable } from '../hooks/useSeekable';
-import { useDebounce } from '../hooks/useDebounce';
+import { useDuration } from '../../hooks/useDuration';
+import { useSeekable } from '../../hooks/useSeekable';
+import { useDebounce } from '../../hooks/useDebounce';
 import { SingleThumbnailView } from './thumbnail/SingleThumbnailView';
 import { useSliderTime } from './useSliderTime';
 import { TestIDs } from '../../utils/TestIDs';
