+++ src/ui/hooks/usePaused.ts
