+++ src/ui/hooks/useThumbnailTrack.ts
